Added department DTO, mapping and changed controller to handle it

+++ Inventory/Server/Controllers/DepartmentController.cs
@@ -1,6 +1,9 @@
 ﻿using Inventory.Server.Services.DepartmentService;
+using Inventory.Server.Resources;
+using AutoMapper;
 using Inventory.Shared;
 using Microsoft.AspNetCore.Mvc;
+using Inventory.Server.Extensions;
 
 namespace Inventory.Server.Controllers
 {
@@ -9,71 +12,79 @@ namespace Inventory.Server.Controllers
     public class DepartmentController : ControllerBase
     {
         private readonly IDepartmentService _departmentService;
-        public DepartmentController(IDepartmentService departmentService)
+        private readonly IMapper _mapper;
+
+        public DepartmentController(IDepartmentService departmentService, IMapper mapper)
         {
             _departmentService = departmentService;
+            _mapper = mapper;
         }
 
         // GET: api/Department
         [HttpGet]
-        public async Task<ActionResult<ServiceResponse<List<Department>>>> GetDepartments()
+        public async Task<ActionResult<ServiceResponse<List<DepartmentResource>>>> GetDepartments()
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState.GetErrorMessages());
             var result = await _departmentService.GetDepartments();
-            return Ok(result);
+            if (!result.Success)
+                return BadRequest(result.Message);
+            var newResult = _mapper.Map<ServiceResponse<List<Department>>, ServiceResponse<List<DepartmentResource>>>(result);
+            return Ok(newResult);
         }
 
         // GET api/Department/5
         [HttpGet("{departmentId}")]
-        public async Task<ActionResult<ServiceResponse<Department>>> GetDepartment(int departmentId)
+        public async Task<ActionResult<ServiceResponse<DepartmentResource>>> GetDepartment(int departmentId)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState.GetErrorMessages());
             var result = await _departmentService.GetDepartment(departmentId);
-            return Ok(result);
+            if (!result.Success)
+                return BadRequest(result.Message);
+            var newResult = _mapper.Map<ServiceResponse<Department>, ServiceResponse<DepartmentResource>>(result);
+            return Ok(newResult);
         }
 
         // POST api/Department
         [HttpPost]
-        public async Task<ActionResult<ServiceResponse<Department>>> CreateDepartment(Department department)
+        public async Task<ActionResult<ServiceResponse<DepartmentResource>>> CreateDepartment([FromBody] SaveDepartmentResource resource)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState.GetErrorMessages());
+            var department = _mapper.Map<SaveDepartmentResource, Department>(resource);
             var result = await _departmentService.CreateDepartment(department);
-            if (result.Success == true)
-            {
-                return Ok(result);
-            }
-            else
-            {
-                return NotFound(result);
-            }
-
+            if (!result.Success)
+                return BadRequest(result.Message);
+            var newResult = _mapper.Map<ServiceResponse<Department>, ServiceResponse<DepartmentResource>>(result);
+            return Ok(newResult);
         }
 
         // PUT api/Department/5
         [HttpPut("{departmentId}")]
-        public async Task<ActionResult<ServiceResponse<Department>>> PutDepartment(int departmentId, [FromBody] Department department)
+        public async Task<ActionResult<ServiceResponse<DepartmentResource>>> PutDepartment(int departmentId, [FromBody] SaveDepartmentResource resource)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState.GetErrorMessages());
+            var department = _mapper.Map<SaveDepartmentResource, Department>(resource);
             var result = await _departmentService.PutDepartment(departmentId, department);
-            if (result.Success == true)
-            {
-                return Ok(result);
-            }
-            else
-            {
-                return NotFound(result);
-            }
+            if (!result.Success)
+                return BadRequest(result.Message);
+            var newResult = _mapper.Map<ServiceResponse<Department>, ServiceResponse<DepartmentResource>>(result);
+            return Ok(newResult);
         }
 
         // DELETE api/Department/5
         [HttpDelete("{departmentId}")]
-        public async Task<ActionResult<ServiceResponse<Department>>> DeleteDepartment(int departmentId)
+        public async Task<ActionResult<ServiceResponse<DepartmentResource>>> DeleteDepartment(int departmentId)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState.GetErrorMessages());
             var result = await _departmentService.DeleteDepartment(departmentId);
-            if (result.Success == true)
-            {
-                return Ok(result);
-            }
-            else
-            {
-                return NotFound(result);
-            }
+            if (!result.Success)
+                return BadRequest(result.Message);
+            var newResult = _mapper.Map<ServiceResponse<Department>, ServiceResponse<DepartmentResource>>(result);
+            return Ok(newResult);
         }
     }
 }
